added speakersList to conferences display + started feedback page on frontend

+++ backend/backend/src/Charisma.Common.Domain/Dtos/Commands/Conferences/SaveConference.cs
@@ -39,7 +39,7 @@ namespace Charisma.Common.Domain.Dtos.Commands.Conferences
         public int CountryId { get; init; }
         public int CountyId { get; init; }
         public int CityId { get; init; }
-        public int Latitude { get; init; }
-        public int Longitude { get; init; }
+        public decimal Latitude { get; init; }
+        public decimal Longitude { get; init; }
     }
 }
